crm-watcher/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
easyocr==1.7.2
filelock==3.19.1
fsspec==2025.9.0
greenlet==3.2.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
ninja==1.13.0
numpy==2.0.2
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
playwright==1.55.0
pyclipper==1.3.0.post6
pyee==13.0.0
python-bidi==0.6.6
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.7
sympy==1.14.0
tifffile==2024.8.30
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
urllib3==2.5.0
